[AS7-5273] Remove dependency on OSGi from Naming/Transaction subsystem
