+++ osgi/integration/pom.xml
@@ -29,19 +29,47 @@
 
     <parent>
         <groupId>org.jboss.as</groupId>
-        <artifactId>jboss-as-parent</artifactId>
+        <artifactId>jboss-as-osgi</artifactId>
         <version>7.2.0.Alpha1-SNAPSHOT</version>
     </parent>
 
-    <groupId>org.jboss.as</groupId>
-    <artifactId>jboss-as-osgi</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>jboss-as-osgi-integration</artifactId>
+    <name>JBoss Application Server: OSGi Integration</name>
 
-    <name>JBoss Application Server: OSGi</name>
+    <dependencies>
+        <dependency>
+            <groupId>org.jboss.as</groupId>
+            <artifactId>jboss-as-web</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss.as</groupId>
+            <artifactId>jboss-as-jmx</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss.as</groupId>
+            <artifactId>jboss-as-naming</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss.as</groupId>
+            <artifactId>jboss-as-transactions</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss.as</groupId>
+            <artifactId>jboss-as-osgi-service</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss.osgi.framework</groupId>
+            <artifactId>jbosgi-framework-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.osgi</groupId>
+            <artifactId>org.osgi.enterprise</artifactId>
+        </dependency>
+    </dependencies>
 
-    <modules>
-        <module>configadmin</module>
-        <module>service</module>
-        <module>integration</module>
-    </modules>
+    <!-- Build -->
+    <build>
+        <plugins>
+        </plugins>
+    </build>
 </project>
