+++ osgi/integration/src/main/java/org/jboss/as/osgi/httpservice/DefaultServletConfig.java
@@ -1,0 +1,69 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2010, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+package org.jboss.as.osgi.httpservice;
+
+import java.util.Dictionary;
+import java.util.Enumeration;
+import java.util.Hashtable;
+
+import javax.servlet.ServletConfig;
+import javax.servlet.ServletContext;
+
+import org.osgi.service.http.HttpContext;
+
+/**
+ * The default {@link HttpContext}
+ *
+ * @author Thomas.Diesler@jboss.com
+ * @since 19-Jul-2012
+ */
+final class DefaultServletConfig implements ServletConfig {
+
+    private final String servletName;
+    private final ServletContext servletContext;
+    private final Dictionary<String, String> initParams = new Hashtable<String, String>();
+
+    DefaultServletConfig(String servletName, ServletContext servletContext) {
+        this.servletName = servletName;
+        this.servletContext = servletContext;
+    }
+
+    @Override
+    public String getServletName() {
+        return servletName;
+    }
+
+    @Override
+    public ServletContext getServletContext() {
+        return servletContext;
+    }
+
+    @Override
+    public String getInitParameter(String name) {
+        return initParams.get(name);
+    }
+
+    @Override
+    public Enumeration<String> getInitParameterNames() {
+        return initParams.keys();
+    }
+}
