+++ osgi/integration/src/main/java/org/jboss/as/osgi/web/WabServletContextFactoryProcessor.java
