+++ osgi/integration/src/main/java/org/jboss/as/osgi/web/WabServletContextServiceProcessor.java
