+++ osgi/integration/src/main/java/org/jboss/as/osgi/web/WebContextActivationProcessor.java
