+++ osgi/service/src/main/java/org/jboss/as/osgi/AbstractSubsystemExtension.java
@@ -1,0 +1,59 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2010, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+
+package org.jboss.as.osgi;
+
+import java.util.List;
+
+import org.jboss.as.controller.OperationContext;
+import org.jboss.as.controller.ServiceVerificationHandler;
+import org.jboss.dmr.ModelNode;
+import org.jboss.msc.service.ServiceBuilder;
+import org.jboss.msc.service.ServiceController;
+import org.jboss.msc.service.StartContext;
+import org.jboss.msc.service.StopContext;
+import org.osgi.framework.BundleContext;
+
+/**
+ * An abstract extension to the OSGi subsystem
+ *
+ * @author thomas.diesler@jboss.com
+ * @since 31-Jul-2012
+ */
+public abstract class AbstractSubsystemExtension implements SubsystemExtension {
+
+    @Override
+    public void performBoottime(OperationContext context, ModelNode operation, ModelNode model, ServiceVerificationHandler verificationHandler, List<ServiceController<?>> newControllers) {
+    }
+
+    @Override
+    public void configureSystemServiceDependencies(ServiceBuilder<?> builder) {
+    }
+
+    @Override
+    public void startSystemServices(StartContext startContext, BundleContext systemContext) {
+    }
+
+    @Override
+    public void stopSystemServices(StopContext stopContext, BundleContext systemContext) {
+    }
+}
