--- osgi/web/src/main/java/org/jboss/as/osgi/web/httpservice/DefaultHttpContext.java
@@ -1,61 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2010, Red Hat, Inc., and individual contributors
- * as indicated by the @author tags. See the copyright.txt file in the
- * distribution for a full listing of individual contributors.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-package org.jboss.as.osgi.web.httpservice;
-
-import java.io.IOException;
-import java.net.URL;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.osgi.framework.Bundle;
-import org.osgi.service.http.HttpContext;
-
-/**
- * The default {@link HttpContext}
- *
- * @author Thomas.Diesler@jboss.com
- * @since 19-Jul-2012
- */
-final class DefaultHttpContext implements HttpContext {
-
-    private final Bundle bundle;
-
-    DefaultHttpContext(Bundle bundle) {
-        this.bundle = bundle;
-    }
-
-    @Override
-    public boolean handleSecurity(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        return true;
-    }
-
-    @Override
-    public URL getResource(String name) {
-        return bundle.getResource(name);
-    }
-
-    @Override
-    public String getMimeType(String name) {
-        return null;
-    }
-}
