+++ pom.xml
@@ -1057,7 +1057,7 @@
 
             <dependency>
                 <groupId>org.jboss.as</groupId>
-                <artifactId>jboss-as-osgi-web</artifactId>
+                <artifactId>jboss-as-osgi-integration</artifactId>
                 <version>${project.version}</version>
             </dependency>
